refactor: use getScoresDirect to fech getVp score
